refactor: separate changelog run concerns
